+++ lib/reveal-ck/commands.rb
@@ -1,0 +1,8 @@
+require_relative 'commands/generate'
+
+require_relative 'commands/print_banner'
+require_relative 'commands/listen_to_rebuild_slides'
+require_relative 'commands/listen_to_reload_browser'
+require_relative 'commands/start_web_server'
+require_relative 'commands/serve_ui'
+require_relative 'commands/serve'